requirements.txt
pip install django
pillow
django-contrib-comments